refactor: made a new component for landing page
